requirements.txt
ccxt==4.3.20
anthropic==0.28.0
fastapi==0.111.0
uvicorn[standard]==0.30.1
APScheduler==3.10.4
streamlit==1.35.0
plotly==5.22.0
python-telegram-bot==21.3
python-dotenv==1.0.1
pydantic==2.7.1
rich==13.7.1
aiosqlite==0.20.0
requests==2.32.3
pandas==2.2.2
